Remove newly introduced `willUpdateManifest` method, for now.
Removes the `willUpdateManifest` method which was originally introduced in
https://github.com/apollographql/apollo-platform-commercial/pull/158, until
we can come to a better agreement as to how it should work.

While the idea of this method was to give user's control over what to do in
the case of an unavailable manifest, the implementation exposes the exact
format of the manifest to the user, preventing us from iterating on it.

It also leaves the exercise of generating and updating the manifest to the
user, which honestly, if documented using this approach would potentially
introduce many additional complexities to deployment processes and severely
undermine the value of the feature.

Some things which we may want to introduce first are:

- The ability to fail-safe or fail-secure on server startup when the
  manifest is not available.
- The ability to fail-safe or fail-secure after a specified interval of
  manifest unavailability.  For example, when the manifest cannot be fetched
  for a specific time, switch to a particular mode of operation.
- A CDN for our manifests to reduce the likelihood of the manifest being
  unavailable.
- Since this extends to other artifacts which are hosted in our cloud,
  potentially, a more blanket recommendation that consumers setup an
  on-premise artifact cache and potentially use a Varnish (or other) proxy
  as an in-between store for artifacts that are hosted in our
  cloud.

As it stands right now though, there is a lack of agreement on a best
practice here that doesn't heavily rely on being locked into our manifest
format.  Further, as far as I'm aware, nobody is actually implementing this
method and no customer (other than "us") has specifically asked for this
ability (even if we _think_ it might be what they need).