Change the `ioredis` test mock to not expire records lacking a `ttl`.
Most cache stores, including Redis, support the notion of not applying a TTL
specifier on a cache record, thus allowing it to survive as long as
possible.

Of course, other factors may cause the record to be evicted (e.g. out of
memory on that node, need to make room for objects which are hotter, etc.),
those decisions are ultimately left up to the controller of those caches.

This paves the way for the next commits.