Preserve other non-`cache` properties on this APQ cache configuration.
This was already the case prior to d9cfe3e04e32b1041e99d8dfd218caa82f7cddd5,
however the rejiggering in that PR removed that functionality, causing any
other properties present on the cache configuration to be dropped.

The original implementation can be seen here:

https://github.com/apollographql/apollo-server/commit/d9cfe3e0#diff-3467ccc78fbd796513acd747ff2110a5L212-L216

This was frustrating to sort out!