Allow `KeyValueCache` implementations to not specify a TTL by using `null`.
Rather than defaulting to a behavior of not applying a TTL unless desired, all
of the current cache implementations specify a default TTL of 300.

This change introduces a specific notion of an intent to not apply a TTL
(and instead leave the eviction decision to the cache controller) by
introducing a concept of `null` for the TTL which will allow individual
subclasses of `KeyValueCache` to react accordingly.  For example, in
Memcached, this requires setting the TTL to 0, but in Redis, it requires
omitting arguments entirely.

This preserves backwards compatible behavior for existing cache
configurations by levering this new `null`ability.