typings: Accept `null` for DataLoader return types in RedisCache.
I encountered this when trying to land #3553, which updates to DataLoader v2.

These type errors were previous masked by a scope problem which resulted in
`this.loader` being `any`-typed inside the callback that's passed to
`DataLoader`'s first parameter (`batchLoadFn`).

Fixing that by leveraging the locally scoped variable revealed that the type
for the loader instance itself is incorrect since `mget` will return `null`
for a key if it is not present in the cache.

This adjusts the typings to fix both of those.