types: Optionality of `cache` on `PersistedQueryOptions`.
The `cache` value is not required to be passed inside of `persistedQueries`
(the top-level object that's passed to Apollo Server's constructor options
to configure the APQ cache).  When it's omitted, it defaults to the (global)
`cache` also passed at the very top-level of the same constructor options.

This wasn't usually a problem in the past, but now that there are other
adjacent properties (like the `ttl` property introduced to control the
default TTL for APQ; https://github.com/apollographql/apollo-server/pull/3623)
this has made it desirable to drop off the `cache` entirely and only pass
`ttl`.