Kinda revert "Decompose `GatewayConfigBase` into types in prep..."
This reverts the decomposition introduced by commit
7398b3929df9211830a783022c4f08ea6875de4f, but also maintains the bit that
was added right after it in ca10ed82ab061c57e0abad.

Specifically, I misjudged with the direction I was headed.  While my intent
was to have `debug` and `logger` be mutually exclusive, it turns out that in
order to preserve the existing behavior, I need to make sure that `debug:
true` is the thing that puts the default logger (`loglevel`) into the
appropriate level where it still prints out various legacy messages.

cc @trevor-scheer