Apply counter-intuitive severity to `debugPrintReports` messages.
In terms of verbosity, and as the name of this option suggests, the output
when `debugPrintReports` is enabled is is either an "info" or a "debug"
level message.

However, we are using `warn` here for compatibility reasons since the
`debugPrintReports` flag pre-dated the existence of log-levels and changing
this to also require `debug: true` (in addition to `debugPrintReports`) just
to reach the level of verbosity to produce the output would be a breaking
change.  The "warn" level is on by default.  There is a similar theory and
comment applied below.

cc @trevor-scheer