Add `release:version-bump:server-and-federation` npm script.
This will make sure to bump the releases of both federation and server to
ensure that the version bumps stay aligned during the prerelease, and
avoiding some occasionally bumping to `alpha.2` while others progress to
`alpha.3` (for example).  This is only a concern for documentation (e.g.
CHANGELOG) and human approachability, not for anything technical.

cc @trevor-scheer