Disable Node.js as a requirement for the publishing workflow.
This is a work-around, but should not preclude or prevent Node.js 6 from
working at runtime, since this is a failure in the local repository's
ability to generate the `proto` file for publishing in
`apollo-engine-reporting-protobuf`, not a limitation of the generated file
(which will be published).

Essentially, the release is being prevented since the Node.js workflow is
failing because `@apollo/protobufjs`'s insistence on not having lockstep
versioning or being constrained by a `package-lock.json`.  This results in
an unexpected update to a transitive dependency (`mkdirp`) which is caused
by `mkdirp` recently changing its `latest` tag to point to the `v1.x.x`
release line rather than the `0.5.x` release line it had been using 4 days
ago.

cc @trevor-scheer