chore: Convert `schemaHash` from `string` to a faux-paque type.
By default, TypeScript uses structural typing (as opposed to nominal typing)
Put another way, if it looks like the type and walks like that type, then
TypeScript lets it be a type.

That's often okay, but it leaves a lot to be desired since a `string` of one
type can just be passed in as `string` for that type and TypeScript won't
complain.  Flow offers opaque types which solve this, but TypeScript doesn't
offer this (yet?).  This Faux-paque type can be used to gain nominal-esque
typing, which is incredibly beneficial during re-factors!

For the `schemaHash`, in particular, this is very much a string
representation that serves a very particular purpose.  Passing it
incorrectly somewhere could be problematic, but we can avoid that
(particularly as I embark on some re-factoring with it momentarily), by
typing it as a more-opaque type prior to refactoring.  Such passing around
of strings can be common, for example, in positional parameters of
functions: like a function that receives five strings, but a parameter ends
up being misaligned with its destination.  With structural typing, it's
completely possible to miss that, but `SchemaHash` will _always_ be a
`SchemaHash` with this fauxpaque-typing.

Happy to not land this, but I think it provides some value.  Input
appreciated!