Introduce `BaseContext` and `DefaultContext`.
We use this notation in so many places to represent the context and the
default context that we should stop referring to it as `Record<string,
any>`.

This doesn't go so far as to track all of those down, but this at least
makes the types exist so we can pursuit that as some point.  Luckily, with
structural typing, the use of this or an in-line type elsewhere should
resolve to a match.