apollo-gateway: split secrets, go through CDN (#4080)
This commit changes the URLs for managed federation configuration.
Previously loaded from the GCS buckets directly, the gateway will now
request these resources from the CDNs at these URLs:

* https://storage-secrets.api.apollographql.com
* https://federation.api.apollographql.com

For users that have safelisted or proxied the GCS buckets in their
infrastructure, this is a breaking change. The listed endpoints are
a replacement for the GCS buckets.