types: Improve `Dispatcher`'s `callTargets` and `invokeHookAsync` types.
This came up as a matter of discussion within the referenced [[Comment]] when
reviewing #3988.  The return value of `targets.map` was previously `any[]`'d
which, we would hope, could be improved upon.

I took a shot at improving it and struggled with the `UnwrapPromise`s when
coupled with an expected `Promise` return of `invokeHookAsync`.  Removing
the unwrapping would seem to still get us the correct type (i.e. an array of
the return value of the stingly-typed life-cycle hook, e.g.
`wilSendResponse` or `didEncounterErrors`) at the invocation sites of
`invokeHookAsync` within the request pipeline (e.g. [[Example]]), and seems
to rid us of the `any[]` within the `Dispatcher`.  Perhaps I'm missing
something, but this would seem to work.  I'm perhaps missing some unexpected
conflict with the `ValueOrPromise` return types from the
`GraphQLRequestListener` methods, but they all return `ValueOrPromise<void>`
right now any time that `invokeHookAsync` is called.

[Comment]: https://github.com/apollographql/apollo-server/pull/3988/files#r421665333
[Example]: https://github.com/apollographql/apollo-server/blob/2284ea06/packages/apollo-server-core/src/requestPipeline.ts#L591-L594