apollo-engine-reporting: fix TypeError on sendReport("unknown")
EngineReportingAgent had three different fields all of which were objects keyed
by executableSchemaId. Their key sets were identical but this was not
represented explicitly in the code. In addition, their TypeScript typings were
of the form `{[x: string]: Foo}`, which means TypeScript will inaccurately
assume that looking up *any* key on the object will return an actual Foo and not
undefined.

This refactor merges the three maps into a single map whose value is a
ReportData with three fields. Additionally, it fixes the type to reflect that
not every executableSchemaId is always in the map.

Mostly this just leads to code whose properties (an executableSchemaId may or
may not have these three pieces of data associated with them, and if they do
then all three are there) can be checked at compile time rather than runtime.

This does actually fix a minor bug: if `sendReport` is called manually with an
executableSchemaId which is not associated with a trace that has been reported
yet (perhaps because it is called before any traces are captured?), the code
used to crash when it tried to evaluate `report.tracesPerQuery`, but now it will
create an empty report for that id and not crash (but still not send it because
it is empty). We do actually document that you can call `sendReport` yourself so
this seems reasonable.