Revert "Release"
This reverts commit f88bc3d86baa0814a85b4483d5417e96c7e5236b.

As noted in the commit prior to this one: 0e8871f, TS errors were
introduced by cherry-picking 4e80d3b (which was targeting a
different branch previously, an easy mistake to make).

Associated git tags were removed from the existing release commit
with the intention that they'll be used in the release commit
following this one.