feat(federation): Use extensions instead of custom `federation` objects (#4313)
Note: BREAKING CHANGE
In order to guarantee the safe usage of extensions, we must now
require `graphql@>14.5.0`. As such, we've narrowed the range in
federation and gateway's `peerDependencies`.

Composition in its current form modifies the graphql-js types in
order to accommodate some additional metadata required during
composition, validation, query planning, and execution. This is
no longer required due to the addition of 'extensions', which
this is intended to be used exactly for.

This commit transitions our usages of these custom metadata
objects into the extensions object where they belong.