Disable file-upload tests on Node.js 14.
Node.js 14 is not LTS yet, but we want to make sure that we're preparing
for and breaking changes that it brings.

We can't support uploads on Node.js 14 before we upgrade to `graphql-upload`
version 10, which is a major breaking change for current upload users.
We're very aware (via issues like
https://github.com/apollographql/apollo-server/issues/3508) that we won't
support it on Node.js 14 without updating it to a newer version.

In a conflicting vein though, we're actually removing `graphql-upload` from
Apollo Server 3.x (and letting the package be usable by itself), so this
just allows us to test the rest of our stack which we intend to support on
Node.js 14, while recognizing that we're not intending on supporting uploads
as a built-in.