fix(federation): complex `@key`s result in unparseable CSDL (#4490)
This fix targets a bug in printComposedSdl where @key directive fields args
weren't being printed with reduced whitespace as @requires and @provides are,
resulting in an invalid / unparseable graphql document.

To reproduce this fix, I first introduced a nested @key into our testing fixtures,
which we should have always had at least one of to begin with. The resulting bug
can be seen in the second commit, and then resolved in the third commit.

I've also added a test to printComposedSdl to ensure the output is parseable.